+++ src/main/scala/au/com/agl/arc/api/API.scala
@@ -166,7 +166,7 @@ object API {
 
   case class DelimitedExtract(name: String, cols: List[ExtractColumn], outputView: String, input: Either[String, URI], settings: Delimited, authentication: Option[Authentication], params: Map[String, String], persist: Boolean, numPartitions: Option[Int], partitionBy: List[String], contiguousIndex: Option[Boolean]) extends ColumnarExtract { val getType = "DelimitedExtract" }
 
-  case class HTTPExtract(name: String, uri: URI, headers: Map[String, String], validStatusCodes: Option[List[Int]], outputView: String, params: Map[String, String], persist: Boolean, numPartitions: Option[Int], partitionBy: List[String]) extends Extract { val getType = "HTTPExtract" }
+  case class HTTPExtract(name: String, uri: URI, method: Option[String], headers: Map[String, String], body: Option[String], validStatusCodes: Option[List[Int]], outputView: String, params: Map[String, String], persist: Boolean, numPartitions: Option[Int], partitionBy: List[String]) extends Extract { val getType = "HTTPExtract" }
 
   case class JDBCExtract(name: String, outputView: String, jdbcURL: String, tableName: String, numPartitions: Option[Int], fetchsize: Option[Int], customSchema: Option[String], driver: java.sql.Driver, params: Map[String, String], persist: Boolean, partitionBy: List[String]) extends Extract { val getType = "JDBCExtract" }