+++ src/test/scala/au/com/agl/arc/extract/HTTPExtractSuite.scala
@@ -10,6 +10,8 @@ import org.mortbay.jetty.handler.{AbstractHandler, ContextHandler, ContextHandle
 import org.mortbay.jetty.{Server, Request, HttpConnection}
 import javax.servlet.http.{HttpServletRequest, HttpServletResponse}
 
+import scala.io.Source
+
 import org.apache.spark.sql._
 import org.apache.spark.sql.functions._
 
@@ -21,23 +23,63 @@ import au.com.agl.arc.util._
 
 class HTTPExtractSuite extends FunSuite with BeforeAndAfter {
 
-  class DataHandler extends AbstractHandler {
+  class GetDataHandler extends AbstractHandler {
     val payload = """[{"booleanDatum":true,"dateDatum":"2016-12-18","decimalDatum":54.321000000000000000,"doubleDatum":42.4242,"integerDatum":17,"longDatum":1520828868,"stringDatum":"test,breakdelimiter","timeDatum":"12:34:56","timestampDatum":"2017-12-20T21:46:54.000Z"},
       |{"booleanDatum":false,"dateDatum":"2016-12-19","decimalDatum":12.345000000000000000,"doubleDatum":21.2121,"integerDatum":34,"longDatum":1520828123,"stringDatum":"breakdelimiter,test","timeDatum":"23:45:16","timestampDatum":"2017-12-29T17:21:49.000Z"}]""".stripMargin
 
     override def handle(target: String, request: HttpServletRequest, response: HttpServletResponse, dispatch: Int) = {
-      response.setContentType("text/html")
-      response.setStatus(HttpServletResponse.SC_OK)
-      response.getWriter().println(payload)
-      HttpConnection.getCurrentConnection().getRequest().setHandled(true) 
+      if (HttpConnection.getCurrentConnection.getRequest.getMethod == "GET") {
+        response.setContentType("text/html")
+        response.setStatus(HttpServletResponse.SC_OK)
+        response.getWriter().println(payload)
+        HttpConnection.getCurrentConnection.getRequest.setHandled(true) 
+      } else {
+        response.setStatus(HttpServletResponse.SC_FORBIDDEN)
+        HttpConnection.getCurrentConnection.getRequest.setHandled(true) 
+      }
     }
   } 
+
+  class PostDataHandler extends AbstractHandler {
+    val payload = """[{"booleanDatum":true,"dateDatum":"2016-12-18","decimalDatum":54.321000000000000000,"doubleDatum":42.4242,"integerDatum":17,"longDatum":1520828868,"stringDatum":"test,breakdelimiter","timeDatum":"12:34:56","timestampDatum":"2017-12-20T21:46:54.000Z"},
+      |{"booleanDatum":false,"dateDatum":"2016-12-19","decimalDatum":12.345000000000000000,"doubleDatum":21.2121,"integerDatum":34,"longDatum":1520828123,"stringDatum":"breakdelimiter,test","timeDatum":"23:45:16","timestampDatum":"2017-12-29T17:21:49.000Z"}]""".stripMargin
+
+    override def handle(target: String, request: HttpServletRequest, response: HttpServletResponse, dispatch: Int) = {
+      if (HttpConnection.getCurrentConnection.getRequest.getMethod == "POST") {
+        response.setContentType("text/html")
+        response.setStatus(HttpServletResponse.SC_OK)
+        response.getWriter().println(payload)
+        HttpConnection.getCurrentConnection.getRequest.setHandled(true) 
+      } else {
+        response.setStatus(HttpServletResponse.SC_FORBIDDEN)
+        HttpConnection.getCurrentConnection.getRequest.setHandled(true) 
+      }
+    }
+  }   
+
+  class PostPayloadHandler extends AbstractHandler {
+    override def handle(target: String, request: HttpServletRequest, response: HttpServletResponse, dispatch: Int) = {
+      if (HttpConnection.getCurrentConnection.getRequest.getMethod == "POST" ) {
+        val inputStream = HttpConnection.getCurrentConnection.getRequest.getInputStream
+        if (Source.fromInputStream(inputStream).mkString == body) {
+          response.setStatus(HttpServletResponse.SC_OK)
+          HttpConnection.getCurrentConnection.getRequest.setHandled(true) 
+        } else {
+          response.setStatus(HttpServletResponse.SC_FORBIDDEN)
+          HttpConnection.getCurrentConnection.getRequest.setHandled(true) 
+        }
+      } else {
+        response.setStatus(HttpServletResponse.SC_FORBIDDEN)
+        HttpConnection.getCurrentConnection.getRequest.setHandled(true) 
+      }
+    }
+  }   
 
   class EmptyHandler extends AbstractHandler {
     override def handle(target: String, request: HttpServletRequest, response: HttpServletResponse, dispatch: Int) = {
       response.setContentType("text/html")
       response.setStatus(HttpServletResponse.SC_OK)
-      HttpConnection.getCurrentConnection().getRequest().setHandled(true) 
+      HttpConnection.getCurrentConnection.getRequest.setHandled(true) 
     }
   }   
 
@@ -47,6 +89,11 @@ class HTTPExtractSuite extends FunSuite with BeforeAndAfter {
   val outputView = "dataset"
   val uri = s"http://localhost:${port}"
   val badUri = s"http://localhost:${port+1}"
+  val get = "get"
+  val post = "post"
+  val payload = "payload"
+  val empty = "empty"
+  val body = "testpayload"
 
   before {
     implicit val spark = SparkSession
@@ -64,12 +111,20 @@ class HTTPExtractSuite extends FunSuite with BeforeAndAfter {
     spark.conf.set("spark.sql.session.timeZone", "UTC")    
 
     // register handlers
-    val dataContext = new ContextHandler("/data");
-    dataContext.setHandler(new DataHandler)
-    val emptyContext = new ContextHandler("/empty");
+    val getContext = new ContextHandler(s"/${get}")
+    getContext.setAllowNullPathInfo(false)
+    getContext.setHandler(new GetDataHandler)
+    val postContext = new ContextHandler(s"/${post}")
+    postContext.setAllowNullPathInfo(false) 
+    postContext.setHandler(new PostDataHandler)    
+    val postPayloadContext = new ContextHandler(s"/${payload}")
+    postPayloadContext.setAllowNullPathInfo(false)   
+    postPayloadContext.setHandler(new PostPayloadHandler)   
+    val emptyContext = new ContextHandler(s"/${empty}")
+    emptyContext.setAllowNullPathInfo(false)
     emptyContext.setHandler(new EmptyHandler)
     val contexts = new ContextHandlerCollection()
-    contexts.setHandlers(Array(dataContext, emptyContext));
+    contexts.setHandlers(Array(getContext, postContext, postPayloadContext, emptyContext));
     server.setHandler(contexts)
 
     // start http server
@@ -85,7 +140,7 @@ class HTTPExtractSuite extends FunSuite with BeforeAndAfter {
     }
   }
 
-  test("HTTPExtract: Can read data") {
+  test("HTTPExtract: Can read data (GET)") {
     implicit val spark = session
     import spark.implicits._
     implicit val logger = LoggerFactory.getLogger(spark.sparkContext.applicationId)
@@ -93,14 +148,16 @@ class HTTPExtractSuite extends FunSuite with BeforeAndAfter {
     val extractDataset = extract.HTTPExtract.extract(
       HTTPExtract(
         name=outputView,
-        uri=new URI(s"${uri}/data"),
+        uri=new URI(s"${uri}/${get}/"),
         headers=Map.empty,
         validStatusCodes=None,
         outputView=outputView,
         params=Map.empty,
         persist=false,
         numPartitions=None,
-        partitionBy=Nil
+        partitionBy=Nil,
+        method=Option("GET"),
+        body=None
       )
     )
 
@@ -125,6 +182,70 @@ class HTTPExtractSuite extends FunSuite with BeforeAndAfter {
     assert(expected.except(actual).count === 0)
   }    
 
+  test("HTTPExtract: Can read data (POST)") {
+    implicit val spark = session
+    import spark.implicits._
+    implicit val logger = LoggerFactory.getLogger(spark.sparkContext.applicationId)
+
+    val extractDataset = extract.HTTPExtract.extract(
+      HTTPExtract(
+        name=outputView,
+        uri=new URI(s"${uri}/${post}/"),
+        headers=Map.empty,
+        validStatusCodes=None,
+        outputView=outputView,
+        params=Map.empty,
+        persist=false,
+        numPartitions=None,
+        partitionBy=Nil,
+        method=Option("POST"),
+        body=None
+      )
+    )
+
+    val actual = spark.read.option("multiLine", "true").json(extractDataset.as[String])
+
+    // JSON does not have DecimalType or TimestampType
+    // JSON will silently drop NullType on write
+    val expected = TestDataUtils.getKnownDataset
+      .withColumn("decimalDatum", $"decimalDatum".cast("double"))
+      .withColumn("timestampDatum", from_unixtime(unix_timestamp($"timestampDatum"), "yyyy-MM-dd'T'HH:mm:ss.SSSXXX"))
+      .drop($"nullDatum")
+
+    val actualExceptExpectedCount = actual.except(expected).count
+    val expectedExceptActualCount = expected.except(actual).count
+    if (actualExceptExpectedCount != 0 || expectedExceptActualCount != 0) {
+      println("actual")
+      actual.show(false)
+      println("expected")
+      expected.show(false)  
+    }
+    assert(actual.except(expected).count === 0)
+    assert(expected.except(actual).count === 0)
+  }  
+
+  test("HTTPExtract: Can post data (POST)") {
+    implicit val spark = session
+    import spark.implicits._
+    implicit val logger = LoggerFactory.getLogger(spark.sparkContext.applicationId)
+
+    val extractDataset = extract.HTTPExtract.extract(
+      HTTPExtract(
+        name=outputView,
+        uri=new URI(s"${uri}/${payload}/"),
+        headers=Map.empty,
+        validStatusCodes=Option(List(200)),
+        outputView=outputView,
+        params=Map.empty,
+        persist=false,
+        numPartitions=None,
+        partitionBy=Nil,
+        method=Option("POST"),
+        body=Option(body)
+      )
+    )
+  }  
+
   test("HTTPExtract: Can handle empty response") {
     implicit val spark = session
     import spark.implicits._
@@ -133,14 +254,16 @@ class HTTPExtractSuite extends FunSuite with BeforeAndAfter {
     val actual = extract.HTTPExtract.extract(
       HTTPExtract(
         name=outputView,
-        uri=new URI(s"${uri}/empty"),
+        uri=new URI(s"${uri}/${empty}/"),
         headers=Map.empty,
         validStatusCodes=None,
         outputView=outputView,
         params=Map.empty,
         persist=false,
         numPartitions=None,
-        partitionBy=Nil
+        partitionBy=Nil,
+        method=None,
+        body=None   
       )
     )
 
@@ -156,14 +279,16 @@ class HTTPExtractSuite extends FunSuite with BeforeAndAfter {
       extract.HTTPExtract.extract(
         HTTPExtract(
           name=outputView,
-          uri=new URI(s"${uri}/fail"),
+          uri=new URI(s"${uri}/fail/"),
           validStatusCodes=None,
           headers=Map.empty,
           outputView=outputView,
           params=Map.empty,
           persist=false,
           numPartitions=None,
-          partitionBy=Nil
+          partitionBy=Nil,
+          method=None,
+          body=None   
         )
       )
     }
@@ -179,14 +304,16 @@ class HTTPExtractSuite extends FunSuite with BeforeAndAfter {
       extract.HTTPExtract.extract(
         HTTPExtract(
           name=outputView,
-          uri=new URI(s"${uri}/empty"),
+          uri=new URI(s"${uri}/${empty}/"),
           headers=Map.empty,
           validStatusCodes=Option(List(201)),
           outputView=outputView,
           params=Map.empty,
           persist=false,
           numPartitions=None,
-          partitionBy=Nil
+          partitionBy=Nil,
+          method=None,
+          body=None   
         )
       )
     }
@@ -209,7 +336,9 @@ class HTTPExtractSuite extends FunSuite with BeforeAndAfter {
           params=Map.empty,
           persist=false,
           numPartitions=None,
-          partitionBy=Nil
+          partitionBy=Nil,
+          method=None,
+          body=None   
         )
       )
     }